Change some more things
To be honest I'm changing so much at once I don't even know what to put.

Mostly makefile changes, some code.

+++ cpu/ports.c
@@ -1,11 +1,11 @@
 #include "ports.h"
 
-uint8_t inb(uint8_t port) {
+uint8_t inb(int port) {
   uint8_t in;
   asm volatile("in %%dx, %%al" : "=a" (in) : "d" (port));
   return in;
 }
 
-void outb(uint8_t port, uint8_t byte) {
+void outb(int port, uint8_t byte) {
   asm volatile("out %%al, %%dx" : : "a" (byte), "d" (port));
 }
